refactor(srv): split subagent_watcher.rs into subagent_watcher/ submodules
Tier 2 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
3717-line file bundled lifecycle, a query/naming API, directory scanning, and
jsonl/journal stream processing behind one impl block plus ~465 lines of
free-standing parsing utilities. Split into subagent_watcher/{mod,types,
query,scan,jsonl,parse,tests}.rs along those already-distinguishable
concerns; process_jsonl_change (328 lines) moved whole, not decomposed
internally. Tests kept as one file (not split per-seam) since most share
fixture helpers and several deliberately exercise scan+jsonl+reconcile
together.

Pure relocation, no logic changes. cargo build -p agentmux-srv clean;
cargo test -p agentmux-srv subagent_watcher:: — 52 passed both before and
after, 0 failed.

<!-- agentmux:agent_id=agentx -->
